restructured all classes into their own files. includes: AddressProcessor, KVStore, Source, SourceManager
